# perf: improve CNPJ algorithm performance