The second CNPJ table can be completely elided, not requiring the algorithm to iterate over it at all.

Ignoring the last digit on the second CNPJ table we have:

```
S1  = cnpj[0] * 5 + cnpj[1] * 4 ... cnpj[4] * 9 ... cnpj[9] * 2
S2' = cnpj[0] * 6 + cnpj[1] * 5 ... cnpj[4] * 2 ... cnpj[9] * 3
```

The difference between S1 and S2' depends on the difference in weights.

```
S2' - S1 = summation of (cnpj[i] * (t2[i] - t1[i]) ) for i from 0 .. len(t1)
```

Where t1 = CNPJ's first weight table and t2 = CNPJ's second weight table.

Thus we have that S2' = S1 + (delta ⋅ digits).

And then, to get the value of S2 we can just add 2 * cnpj[12] to S2', which is the last element of the second CNPJ table multiplied by the first verifying digit of the CNPJ.

```
S2 = S2' + 2 * cnpj[12]
```

The delta for all indexes is 1, except for index 4, which the delta is -7. So, in the algorithm, we can just hard code a special treatment for index 4.

```
goos: linux
goarch: amd64
pkg: github.com/phenpessoa/br
cpu: Intel(R) Core(TM) Ultra 7 155H
                         │   old.txt    │               new.txt               │
                         │    sec/op    │   sec/op     vs base                │
GenerateCNPJ-21            102.55n ± 1%   77.38n ± 1%  -24.54% (p=0.000 n=10)
CNPJ_IsValid18-21           26.72n ± 2%   15.51n ± 3%  -41.95% (p=0.000 n=10)
CNPJ_IsValid14-21           18.67n ± 1%   17.10n ± 2%   -8.41% (p=0.000 n=10)
CNPJ_IsValid18Invalid-21    26.61n ± 2%   15.20n ± 2%  -42.88% (p=0.000 n=10)
CNPJ_IsValid14Invalid-21    18.62n ± 1%   17.12n ± 1%   -8.03% (p=0.000 n=10)
CNPJ_String18-21            52.16n ± 5%   37.73n ± 8%  -27.67% (p=0.000 n=10)
CNPJ_String14-21            41.90n ± 1%   41.41n ± 1%   -1.16% (p=0.000 n=10)
geomean                     34.26n        26.14n       -23.69%
```